# docs/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: rosmon HTTP API
  version: 1.0.0
  description: |
    REST + Server-Sent Events service di atas RouterOS API. Multi-router:
    semua endpoint (kecuali Device CRUD & Health) berada di bawah
    `/api/v1/devices/{device_id}/...` agar satu instance bisa melayani
    banyak router sekaligus.

    ## Response envelope

    Semua endpoint pakai envelope konsisten:

    - **Success single**: `{ "data": { ... } }`
    - **Success list**:   `{ "data": [ ... ], "meta": { "count": N } }`
    - **Error**:          `{ "error": { "code": "...", "message": "...", "path": "..." } }`

    Error `code` stabil — frontend bisa branch dengan aman. Mapping:

    | Code | HTTP | Trigger |
    |---|---|---|
    | `NOT_FOUND` | 404 | `mikrotik.ErrNotFound` |
    | `INVALID_ARGUMENT` | 400 | `mikrotik.ErrInvalidArgument` |
    | `AMBIGUOUS` | 409 | `mikrotik.ErrAmbiguous` |
    | `VALIDATION` | 400 | body gagal validasi `gin` binding |
    | `TIMEOUT` | 504 | `context.DeadlineExceeded` |
    | `CANCELED` | 499 | `context.Canceled` |
    | `INTERNAL` | 500 | uncategorized / panic recovery |
    | `DEVICE_NOT_FOUND` | 404 | device middleware: device ID tidak terhubung |
    | `SERVICE_UNAVAILABLE` | 503 | InfluxDB / dependency lain down |

    Setiap response menyertakan header `X-Request-ID` (16-char hex) untuk
    korelasi log.

    ## Authentication

    Service ini sekarang mendukung **JWT access + refresh token**
...

...
    4. `POST /auth/logout` revoke refresh token.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /devices/{device_id}/hotspot/profile-configs:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - Profile Configs
      summary: List semua profile config untuk device
      responses:
        '200':
          description: List configs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/ProfileConfigResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '429':
          $ref: '#/components/responses/ErrorRateLimit'
  /devices/{device_id}/hotspot/profile-configs/sync:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    post:
      tags:
        - Profile Configs
      summary: Sync profile router → DB + auto-inject on-login script
      description: |
        Tarik semua hotspot profile dari MikroTik router lalu reconcile dengan
        tabel `hotspot_profile_configs`:

          - Profile router yang **belum ada** di DB → di-insert dengan default
            `expiry_mode: "0"`, `price: 0`. Operator perlu PUT untuk set price.
          - Profile yang **sudah ada** di DB → tidak ditimpa (price/mode
            operator dipertahankan).
          - Config DB yang profile-nya **hilang** dari router → masuk
            `orphan` (tidak dihapus otomatis).
          - Untuk semua profile yang sinkron, on-login script di-push ke
            router (best-effort; per-profile error masuk `inject_failed`).

        Membutuhkan device online. Return 503 kalau router tidak terhubung.
      responses:
        '200':
          description: Sync result
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileConfigSyncResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '503':
          description: Device manager not configured atau router offline
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /devices/{device_id}/hotspot/profile-configs/{profile_name}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - name: profile_name
        in: path
        required: true
        schema:
          type: string
        example: default
    get:
      tags:
        - Profile Configs
      summary: Detail config satu profile
      responses:
        '200':
          description: Profile config
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileConfigResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    put:
      tags:
        - Profile Configs
      summary: Buat / update config (upsert)
      description: |
        Idempotent upsert by (device_id, profile_name). Return 201 kalau record
        baru di-create, 200 kalau update. Validity di-normalize ke RouterOS
        canonical sebelum simpan (mis. `168h` → `1w`).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProfileConfigUpsertRequest'
            examples:
              paket_mingguan:
                summary: Paket 1 minggu
                value:
                  expiry_mode: rem
                  validity: 7d
                  price: 10000
                  sell_price: 12000
                  lock_mac: false
              paket_bulanan_premium:
                summary: Paket bulanan premium (lock MAC)
                value:
                  expiry_mode: remc
                  validity: 30d
                  price: 35000
                  sell_price: 50000
                  lock_mac: true
      responses:
        '200':
          description: Config ter-update
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileConfigResponse'
        '201':
          description: Config baru dibuat
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileConfigResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
    delete:
      tags:
        - Profile Configs
      summary: Hapus config profile
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '401':
          $ref: '#/components/responses/ErrorUnauthorized'
        '403':
          $ref: '#/components/responses/ErrorForbidden'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /devices/{device_id}/hotspot/profiles:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/CacheQuery'
      - $ref: '#/components/parameters/TtlQuery'
    get:
      tags:
        - Hotspot Profiles
      summary: List profiles
      description: List `/ip/hotspot/user/profile`. Cache opsional.
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/HotspotProfileResponse'
        '500':
          $ref: '#/components/responses/ErrorInternal'
    post:
      tags:
        - Hotspot Profiles
      summary: Create profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileCreateRequest'
            examples:
              vip_2m:
                summary: Profile VIP 2M/2M
                value:
                  name: vip-2m
                  address_pool: hs-pool
                  rate_limit: 2M/2M
                  shared_users: 1
                  on_login: ':log info "login $user from $address";'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
...
  /devices/{device_id}/hotspot/profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags:
        - Hotspot Profiles
      summary: Get profile by ID
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/HotspotProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    put:
      tags:
        - Hotspot Profiles
      summary: Update profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileUpdateRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags:
        - Hotspot Profiles
      summary: Delete profile (cascade)
      description: |
        Hapus profile. Kalau body `{name: ...}` di-set, workflow akan cleanup
        scheduler/script yang nama-nya match (cascade delete).
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HotspotProfileDeleteRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
  /devices/{device_id}/ppp/profiles:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
    get:
      tags:
        - PPP
      summary: List PPP profiles
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/PPPProfileResponse'
    post:
      tags:
        - PPP
      summary: Create PPP profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PPPProfileCreateRequest'
      responses:
        '201':
          description: Profile created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/IDResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
...
  /devices/{device_id}/ppp/profiles/{id}:
    parameters:
      - $ref: '#/components/parameters/DeviceIdPath'
      - $ref: '#/components/parameters/IdPath'
    get:
      tags:
        - PPP
      summary: Get PPP profile by ID
      responses:
        '200':
          description: Profile detail
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PPPProfileResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    put:
      tags:
        - PPP
      summary: Update PPP profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PPPProfileUpdateRequest'
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '400':
          $ref: '#/components/responses/ErrorValidation'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
    delete:
      tags:
        - PPP
      summary: Delete PPP profile
      responses:
        '204':
          $ref: '#/components/responses/NoContentResponse'
        '404':
          $ref: '#/components/responses/ErrorNotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token (HS256). Diperoleh via `POST /auth/login`.
        Wajib di-include sebagai header `Authorization: Bearer <token>`
        kecuali untuk endpoint public (login, refresh, logout, /healthz).
  schemas:
    Envelope:
      type: object
      description: Response envelope generik. Field `data` dan `error` saling eksklusif.
      properties:
        data:
          description: Payload sukses. Tipe spesifik tergantung endpoint.
        meta:
          type: object
          additionalProperties: true
          description: Metadata opsional, biasanya berisi `count` untuk list.
          example:
            count: 47
        error:
          $ref: '#/components/schemas/ErrorPayload'
    ErrorEnvelope:
      type: object
      description: Envelope untuk response error (status 4xx/5xx).
      required:
        - error
      properties:
        error:
          $ref: '#/components/schemas/ErrorPayload'
...
    IDResponse:
      type: object
      description: Body untuk endpoint create yang hanya kembalikan `.id`.
      required:
        - id
      properties:
        id:
          type: string
          description: RouterOS internal ID (`.id`).
          example: '*5'
...
    HotspotProfileResponse:
      type: object
      required:
        - id
        - name
        - shared_users
        - add_mac_cookie
        - transparent_proxy
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: default
        address_pool:
          type: string
          example: hs-pool
        rate_limit:
          type: string
          example: 2M/2M
        shared_users:
          type: integer
          minimum: 0
          example: 1
        status_autorefresh:
          type: string
          example: 1m
        on_login:
          type: string
          example: ':log info $user'
        on_logout:
          type: string
        parent_queue:
          type: string
        idle_timeout:
          type: string
          example: 5m
        keepalive_timeout:
          type: string
          example: 2m
        session_timeout:
          type: string
          example: 1h
        mac_cookie_timeout:
          type: string
          example: 3d
        add_mac_cookie:
          type: boolean
        transparent_proxy:
          type: boolean
    HotspotProfileCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: vip-2m
        address_pool:
          type: string
          example: hs-pool
        rate_limit:
          type: string
          example: 2M/2M
        shared_users:
          type: integer
          minimum: 0
          example: 1
        status_autorefresh:
          type: string
          example: 1m
        on_login:
          type: string
        parent_queue:
          type: string
    HotspotProfileUpdateRequest:
      type: object
      description: Sparse update.
      properties:
        name:
          type: string
        address_pool:
          type: string
        rate_limit:
          type: string
        shared_users:
          type: integer
          minimum: 0
          nullable: true
        status_autorefresh:
          type: string
        on_login:
          type: string
          nullable: true
        parent_queue:
          type: string
    HotspotProfileDeleteRequest:
      type: object
      description: |
        Optional body untuk DELETE profile. Kalau `name` di-set, workflow akan
        cleanup scheduler/script yang nama-nya match (cascade delete).
      properties:
        name:
          type: string
...
    PPPProfileResponse:
      type: object
      required:
        - id
        - name
        - disabled
      properties:
        id:
          type: string
          example: '*1'
        name:
          type: string
          example: default
        local_address:
          type: string
        remote_address:
          type: string
        rate_limit:
          type: string
          example: 10M/10M
          description: Format rx/tx.
        session_timeout:
          type: string
          example: 1h30m
          description: RouterOS duration.
        idle_timeout:
          type: string
          example: 10m
        parent_queue:
          type: string
          example: global
        on_up:
          type: string
          description: Script optional dijalankan saat session up.
        on_down:
          type: string
          description: Script optional dijalankan saat session down.
        disabled:
          type: boolean
        comment:
          type: string
    PPPProfileCreateRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 128
          example: vip
        local_address:
          type: string
        remote_address:
          type: string
        rate_limit:
          type: string
          example: 10M/10M
        session_timeout:
          type: string
          example: 1h30m
        idle_timeout:
          type: string
          example: 10m
        parent_queue:
          type: string
        on_up:
          type: string
        on_down:
          type: string
        disabled:
          type: boolean
          nullable: true
        comment:
          type: string
...
    ProfileConfigResponse:
      type: object
      properties:
        id:
          type: integer
          example: 1
        device_id:
          type: integer
          example: 1
        profile_name:
          type: string
          example: default
        expiry_mode:
          type: string
          enum:
            - '0'
            - rem
            - ntf
            - remc
            - ntfc
          description: |
            Mode aksi saat user expired:
            - `0`   = none / disable
            - `rem` = hapus user
            - `ntf` = set limit-uptime ke 0 + kick
            - `remc`= hapus + catat transaksi
            - `ntfc`= notice + catat transaksi
          example: rem
        validity:
          type: string
          description: |
            Format RouterOS canonical (hasil `rosfmt.FormatDuration`). Saat
            request, accept format apapun yang dikenali rosfmt (`168h`, `7d`,
            `1w`, `1mo`, `1y`); server normalize sebelum simpan.
          example: 1w
        price:
          type: integer
          example: 10000
        sell_price:
          type: integer
          example: 12000
        lock_mac:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProfileConfigUpsertRequest:
      type: object
      required:
        - expiry_mode
        - validity
      properties:
        expiry_mode:
          type: string
          enum:
            - '0'
            - rem
            - ntf
            - remc
            - ntfc
        validity:
          type: string
          description: |
            Durasi paket. Accept RouterOS (`7d`, `1w`, `1mo`, `1y`) atau Go
            (`168h`). Server normalize ke canonical RouterOS format saat simpan.
          example: 7d
        price:
          type: integer
          minimum: 0
          example: 10000
        sell_price:
          type: integer
          minimum: 0
          example: 12000
        lock_mac:
          type: boolean
          default: false
    ProfileConfigSyncResponse:
      type: object
      description: |
        Ringkasan hasil sync profile router → DB. Tiap kategori berisi
        profile name (string) yang masuk kategori tsb.
      required:
        - synced
        - created
        - orphan
        - injected
        - inject_failed
      properties:
        synced:
          type: array
          items:
            type: string
          description: |
            Profile router yang sudah ada di DB → tidak ditimpa (operator value
            seperti price/mode dipertahankan).
          example:
            - default
            - 30day
        created:
          type: array
          items:
            type: string
          description: |
            Profile router baru yang di-insert ke DB dengan default
            `expiry_mode: "0"`, `price: 0`, `validity: ""`. Operator perlu PUT
            untuk set price/mode agar selling record aktif.
          example:
            - new-promo
        orphan:
          type: array
          items:
            type: string
          description: |
            Config DB yang profile-nya sudah hilang dari router (mungkin
            di-delete via WinBox/CLI). Tidak dihapus otomatis dari DB.
          example:
            - legacy-profile
        injected:
          type: array
          items:
            type: string
          description: |
            Profile yang on-login script-nya berhasil di-push ke router.
          example:
            - default
            - 30day
        inject_failed:
          type: array
          items:
            type: string
          description: |
            Profile yang gagal di-inject. Format: `<profile>: <error>`. Tidak
            fatal — sync tetap completed dan profile lain tetap di-process.
          example: []
  parameters:
    DeviceIdPath:
      name: device_id
      in: path
      required: true
      description: Numeric device ID.
      schema:
        type: integer
        format: int64
        example: 1
    IdPath:
      name: id
      in: path
      required: true
      description: RouterOS internal ID (`.id`), biasanya format `*N` (mis. `*1`, `*2A`).
      schema:
        type: string
        example: '*1'
...
    CacheQuery:
      name: cache
      in: query
      required: false
      description: Aktifkan caching untuk endpoint GET ini. Default `false`.
      schema:
        type: boolean
        default: false
    TtlQuery:
      name: ttl
      in: query
      required: false
      description: TTL cache dalam format Go duration (mis. `60s`, `5m`). Hanya berlaku saat `cache=true`.
      schema:
        type: string
        pattern: ^[0-9]+(ns|us|µs|ms|s|m|h)$
        example: 60s
...
  responses:
...
    ErrorValidation:
      description: Body request gagal validasi schema (gin binding).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            validation_failed:
              summary: Field wajib tidak ada
              value:
                error:
                  code: VALIDATION
                  message: request validation failed
                  path: /api/v1/devices/rb-main/hotspot/users
                  details: 'Key: ''HotspotUserCreateRequest.Name'' Error:Field validation for ''Name'' failed on the ''required'' tag'
    ErrorNotFound:
      description: Resource yang dirujuk tidak ditemukan.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            user_not_found:
              summary: User tidak ada
              value:
                error:
                  code: NOT_FOUND
                  message: 'mikrotik: not found'
                  path: /api/v1/devices/rb-main/hotspot/users/*99
...
    ErrorInternal:
      description: Error internal server. Dipakai untuk uncategorized + panic recovery.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            internal:
              summary: Unexpected error
              value:
                error:
                  code: INTERNAL
                  message: unexpected error during operation
                  path: /api/v1/devices/rb-main/hotspot/users
...
    ErrorUnauthorized:
      description: |
        Token JWT tidak ada / invalid / expired / wrong-type. Response menyertakan
        header `WWW-Authenticate: Bearer realm="api"`.
      headers:
        WWW-Authenticate:
          description: Bearer challenge
          schema:
            type: string
            example: Bearer realm="api"
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            missing:
              summary: Header Authorization tidak ada
              value:
                error:
                  code: UNAUTHORIZED
                  message: missing or malformed Authorization header
                  path: /api/v1/devices
            expired:
              summary: Access token expired
              value:
                error:
                  code: TOKEN_EXPIRED
                  message: token expired
                  path: /api/v1/devices
    ErrorForbidden:
      description: |
        Token valid tapi role tidak diizinkan mengakses endpoint ini
        (mis. operator coba akses /auth/users yang admin-only).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            role_forbidden:
              summary: Role tidak cukup
              value:
                error:
                  code: FORBIDDEN
                  message: role not authorized for this action
                  path: /api/v1/auth/users
    ErrorRateLimit:
      description: |
        Terlalu banyak request. Response menyertakan header `Retry-After`
        (detik). Berlaku untuk per-user (REST), per-IP (anon endpoint),
        per-endpoint-heavy (voucher generate, bulk-delete, reboot, shutdown),
        dan SSE subscriber cap (per-topic / per-device).
      headers:
        Retry-After:
          description: Detik sebelum boleh retry
          schema:
            type: integer
            example: 30
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          examples:
            per_user:
              summary: Per-user limit terlewat
              value:
                error:
                  code: RATE_LIMIT
                  message: too many requests, slow down
                  path: /api/v1/devices
                  details:
                    retry_after_s: 30
                    scope: user
            sse_topic_cap:
              summary: SSE per-topic subscriber cap terlewat
              value:
                error:
                  code: RATE_LIMIT
                  message: sse subscriber cap exceeded
                  path: /api/v1/devices/1/stream/hotspot/active
                  details:
                    retry_after_s: 30
                    scope: topic
    NoContentResponse:
      description: Operasi sukses tanpa body response (204 No Content).